extractor/requirements.txt
flask
redis
psycopg2-binary
bleak

